# Faster colorEnabled() evaluation by reading env capability on startup