This speeds up logging when when color is enabled.

`colorEnabled()` is called every time we log, to determine whether to apply ANSI escape sequences to color log output.

Checking `consoleLogger.ColorEnabled` is fast enough as an exit early, but we were also checking whether the environment is capable of color on each invocation. Now we only do that once, at startup, and use that cached result.

```
14:03 $ benchstat before.out after.out
name                                       old time/op    new time/op    delta
LogColorEnabled/enabled-12                   95.0ns ± 0%     0.6ns ± 0%  -99.39%  (p=0.000 n=5+4)
LogColorEnabled/disabled_console_color-12    0.58ns ± 0%    0.58ns ± 0%     ~     (all equal)
LogColorEnabled/disabled_term_color-12       96.0ns ± 0%     0.6ns ± 0%  -99.40%  (p=0.008 n=5+5)

name                                       old alloc/op   new alloc/op   delta
LogColorEnabled/enabled-12                    0.00B          0.00B          ~     (all equal)
LogColorEnabled/disabled_console_color-12     0.00B          0.00B          ~     (all equal)
LogColorEnabled/disabled_term_color-12        0.00B          0.00B          ~     (all equal)

name                                       old allocs/op  new allocs/op  delta
LogColorEnabled/enabled-12                     0.00           0.00          ~     (all equal)
LogColorEnabled/disabled_console_color-12      0.00           0.00          ~     (all equal)
LogColorEnabled/disabled_term_color-12         0.00           0.00          ~     (all equal)
```